// client/src/components/simulators/CardDispenser/src/enums/InputFields.ts
enum InputFields {
  ISSUER_CODE = 'issuerCode',
  DOCUMENT_NR = 'documentNumber',
  DOCUMENT_TYPE = 'documentType',
  NAME_PRIMARY = 'namePrimary',
  NAME_SECONDARY = 'nameSecondary',
  SEX = 'sex',
  NATIONALITY = 'nationality',
  DATE_OF_BIRTH = 'dateOfBirth',
  DATE_OF_EXPIRY = 'dateOfExpiry',
}
export default InputFields;